separate upload initialization from document registration
