+++ python-backend/app/api/mines/permits/permit_amendment/resources/permit_amendment_document.py
@@ -30,15 +30,6 @@ class PermitAmendmentDocumentListResource(Resource, UserMixin):
     parser.add_argument('document_manager_guid', type=str, store_missing=False)
     parser.add_argument('filename', type=str, store_missing=False)
 
-    @requires_role_edit_permit
-    def post(self, mine_guid):
-        mine = Mine.find_by_mine_guid(mine_guid)
-        if not mine:
-            raise NotFound('Mine not found.')
-
-        return DocumentManagerService.initializeFileUploadWithDocumentManager(
-            request, mine, 'permits')
-
     @api.expect(parser)
     @api.marshal_with(PERMIT_AMENDMENT_DOCUMENT_MODEL, code=201)
     @requires_role_edit_permit
